Separate primary_config object from config.
This allows us to remove the Primary's config and the bootstrap object
from aktualizr-secondary.
