+++ src/libaktualizr/CMakeLists.txt
@@ -11,21 +11,27 @@ set(LIBAKTUALIZR_PUBLIC_HEADERS
 
 aktualizr_source_file_checks(${LIBAKTUALIZR_PUBLIC_HEADERS})
 
+# note: the config object is composed with multiple sub-config objects that live
+# close to the modules they correspond to. To make the config module as
+# self-contained as possible, the method definitions of these sub objects are
+# also added to the module with CMake `target_sources(config PRIVATE ...)`
+# declarations.
 # config has to go first, as other libraries append sources to it
-add_subdirectory("config")
+add_library(config OBJECT)
 
-add_subdirectory("utilities")
 add_subdirectory("bootloader")
 add_subdirectory("bootstrap")
 add_subdirectory("campaign")
+add_subdirectory("config")
 add_subdirectory("crypto")
 add_subdirectory("http")
-add_subdirectory("primary")
-add_subdirectory("uptane")
 add_subdirectory("logging")
-add_subdirectory("storage")
 add_subdirectory("package_manager")
+add_subdirectory("primary")
+add_subdirectory("storage")
 add_subdirectory("telemetry")
+add_subdirectory("uptane")
+add_subdirectory("utilities")
 
 # deprecated, we recommend using aktualizr_lib
 add_library(aktualizr_static_lib STATIC
@@ -38,12 +44,13 @@ add_library(aktualizr_static_lib STATIC
     $<TARGET_OBJECTS:crypto>
     $<TARGET_OBJECTS:http>
     $<TARGET_OBJECTS:jsoncpp>
-    $<TARGET_OBJECTS:package_manager>
-    $<TARGET_OBJECTS:primary>
-    $<TARGET_OBJECTS:utilities>
     $<TARGET_OBJECTS:logging>
+    $<TARGET_OBJECTS:package_manager>
+    $<TARGET_OBJECTS:primary_config>
+    $<TARGET_OBJECTS:primary>
     $<TARGET_OBJECTS:storage>
-    $<TARGET_OBJECTS:uptane>)
+    $<TARGET_OBJECTS:uptane>
+    $<TARGET_OBJECTS:utilities>)
 
 add_library(aktualizr_lib SHARED
     $<TARGET_OBJECTS:asn1>
@@ -55,12 +62,14 @@ add_library(aktualizr_lib SHARED
     $<TARGET_OBJECTS:crypto>
     $<TARGET_OBJECTS:http>
     $<TARGET_OBJECTS:jsoncpp>
-    $<TARGET_OBJECTS:package_manager>
-    $<TARGET_OBJECTS:primary>
-    $<TARGET_OBJECTS:utilities>
     $<TARGET_OBJECTS:logging>
+    $<TARGET_OBJECTS:package_manager>
+    $<TARGET_OBJECTS:primary_config>
+    $<TARGET_OBJECTS:primary>
     $<TARGET_OBJECTS:storage>
-    $<TARGET_OBJECTS:uptane>)
+    $<TARGET_OBJECTS:uptane>
+    $<TARGET_OBJECTS:utilities>)
+
 target_link_libraries(aktualizr_lib ${AKTUALIZR_EXTERNAL_LIBS})
 set_target_properties(aktualizr_lib PROPERTIES LIBRARY_OUTPUT_NAME aktualizr)
 install(TARGETS aktualizr_lib LIBRARY DESTINATION lib COMPONENT aktualizr)
